+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # saltstack-age change log
 
+## 0.4.0
+
+* feat: allow configuration of an identity string using the `AGE_IDENTITY`
+  environment variable and the `age_identity` configuration directive
+
 ## 0.3.0
 
 * fix: add support for nested pillar data